// src/app/http/model/sessionCreateInput.ts
/**
 * Tablk
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 0.0.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { ImageBase64 } from "./imageBase64";

export interface SessionCreateInput {
  name: string;
  description?: string;
  catchphrase?: string;
  scenarioName?: string;
  imageBase64?: ImageBase64;
}